Harden net_core capsule paths
Split DHCP, DNS, TCP, UDP, state, runner, and smoltcp device paths into bounded modules with export-only roots.

Reject invalid endpoints, remove stale DNS state on DHCP transitions, fail closed on DNS zero-address answers, and preserve socket-table cleanup on bind/connect failures.
