+++ userland/capsule_net_core/src/device/mod.rs
@@ -14,57 +14,17 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
+pub mod capabilities;
 pub mod mac;
+pub mod nic_device;
+pub mod read_mac;
+pub mod receive;
 pub mod rx;
+pub mod rx_token;
+pub mod transmit;
 pub mod tx;
+pub mod tx_token;
+pub mod types;
 
-use alloc::vec::Vec;
-use smoltcp::phy::{Device, DeviceCapabilities, Medium};
-use smoltcp::time::Instant;
-
-pub struct NicDevice { pub port: u32 }
-pub struct NicRxToken(Vec<u8>);
-pub struct NicTxToken { pub port: u32 }
-
-impl smoltcp::phy::RxToken for NicRxToken {
-    fn consume<R, F>(self, f: F) -> R
-    where F: FnOnce(&mut [u8]) -> R {
-        let mut frame = self.0;
-        f(&mut frame)
-    }
-}
-
-impl smoltcp::phy::TxToken for NicTxToken {
-    fn consume<R, F>(self, len: usize, f: F) -> R
-    where F: FnOnce(&mut [u8]) -> R {
-        let mut buf = alloc::vec![0u8; len];
-        let r = f(&mut buf);
-        tx::send_frame(self.port, &buf);
-        r
-    }
-}
-
-impl Device for NicDevice {
-    type RxToken<'a> = NicRxToken where Self: 'a;
-    type TxToken<'a> = NicTxToken where Self: 'a;
-
-    fn receive(&mut self, _now: Instant) -> Option<(Self::RxToken<'_>, Self::TxToken<'_>)> {
-        let frame = rx::poll_frame(self.port)?;
-        Some((NicRxToken(frame), NicTxToken { port: self.port }))
-    }
-
-    fn transmit(&mut self, _now: Instant) -> Option<Self::TxToken<'_>> {
-        Some(NicTxToken { port: self.port })
-    }
-
-    fn capabilities(&self) -> DeviceCapabilities {
-        let mut c = DeviceCapabilities::default();
-        c.max_transmission_unit = 1514;
-        c.medium = Medium::Ethernet;
-        c
-    }
-}
-
-pub fn mac(port: u32) -> Option<[u8; 6]> {
-    mac::read_mac(port)
-}
+pub use read_mac::mac;
+pub use types::NicDevice;
