+++ userland/capsule_net_core/src/device/read_mac.rs
@@ -14,7 +14,6 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-pub mod dispatch;
-pub mod resolve_a;
-
-pub use dispatch::dispatch;
+pub fn mac(port: u32) -> Option<[u8; 6]> {
+    crate::device::mac::read_mac(port)
+}
