+++ userland/capsule_net_core/src/iface/dhcp/emit_status_selfcheck.rs
@@ -14,7 +14,28 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-pub mod dispatch;
-pub mod resolve_a;
+use nonos_libc::mk_debug;
 
-pub use dispatch::dispatch;
+use crate::iface::dhcp::{write_decimal_u8, write_octet_quad};
+use crate::server::handlers::dhcp_status::encode_body;
+
+pub fn emit_status_selfcheck() {
+    let mut body = [0u8; 18];
+    encode_body(&mut body);
+    let state_code = body[0];
+    let ip = [body[1], body[2], body[3], body[4]];
+    let mut buf = [0u8; 64];
+    let mut pos = 0usize;
+    for &b in b"[NET-CORE] lease-status state=" {
+        buf[pos] = b;
+        pos += 1;
+    }
+    pos = write_decimal_u8::write_decimal_u8(&mut buf, pos, state_code);
+    for &b in b" ip=" {
+        buf[pos] = b;
+        pos += 1;
+    }
+    pos = write_octet_quad::write_octet_quad(&mut buf, pos, ip);
+    buf[pos] = b'\n';
+    mk_debug(buf.as_ptr(), pos + 1);
+}
