+++ userland/capsule_net_core/src/protocol/tcp.rs
@@ -16,9 +16,7 @@
 
 pub const MAGIC_NTCP: u32 = 0x4E54_4350;
 
-pub const OP_LISTEN: u16 = 2;
 pub const OP_CONNECT: u16 = 3;
-pub const OP_ACCEPT: u16 = 4;
 pub const OP_SEND: u16 = 5;
 pub const OP_RECV: u16 = 6;
 pub const OP_CLOSE: u16 = 7;
@@ -30,3 +28,4 @@ pub const E_BAD_LEN: u16 = 4;
 pub const E_NO_SOCKET: u16 = 5;
 pub const E_RX_EMPTY: u16 = 11;
 pub const E_NOT_CONNECTED: u16 = 12;
+pub const E_BAD_ADDR: u16 = 13;
