+++ userland/capsule_net_core/src/server/handlers/tcp/connect/endpoint.rs
@@ -14,7 +14,22 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-pub mod dispatch;
-pub mod resolve_a;
+use smoltcp::wire::{IpAddress, Ipv4Address};
 
-pub use dispatch::dispatch;
+use crate::protocol::tcp::{E_BAD_ADDR, E_BAD_LEN};
+use crate::server::handlers::tcp::connect::types::Endpoint;
+
+pub fn parse(body: &[u8]) -> Result<Endpoint, u16> {
+    if body.len() < 6 {
+        return Err(E_BAD_LEN);
+    }
+    let ip = [body[0], body[1], body[2], body[3]];
+    let port = u16::from_le_bytes([body[4], body[5]]);
+    if ip == [0, 0, 0, 0] || port == 0 {
+        return Err(E_BAD_ADDR);
+    }
+    Ok(Endpoint {
+        remote: IpAddress::Ipv4(Ipv4Address(ip)),
+        port,
+    })
+}
