+++ userland/capsule_net_core/src/server/handlers/tcp/connect/handle.rs
@@ -14,7 +14,19 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-pub mod dispatch;
-pub mod resolve_a;
+use crate::protocol::tcp::{MAGIC_NTCP, OP_CONNECT};
+use crate::server::handlers::tcp::connect::{endpoint, open_socket, reply_outcome};
+use crate::server::parse_req::Request;
+use crate::server::respond::reply;
 
-pub use dispatch::dispatch;
+pub fn handle(sender_pid: u32, req: &Request, body: &[u8], tx: &mut [u8]) {
+    let endpoint = match endpoint::parse(body) {
+        Ok(endpoint) => endpoint,
+        Err(errno) => {
+            let _ = reply(sender_pid, MAGIC_NTCP, OP_CONNECT, errno, req.request_id, &[], tx);
+            return;
+        }
+    };
+    let outcome = open_socket::open_socket(sender_pid, endpoint);
+    reply_outcome::reply_outcome(sender_pid, req.request_id, outcome, tx);
+}
