+++ userland/capsule_net_core/src/server/handlers/tcp/mod.rs
@@ -14,27 +14,12 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-mod close;
-mod connect;
-mod recv;
-mod send;
-mod state;
+pub mod close;
+pub mod connect;
+pub mod dispatch;
+pub mod recv;
+pub mod send;
+pub mod smoltcp_state_to_code;
+pub mod state;
 
-use crate::protocol::tcp::{
-    E_BAD_OP, MAGIC_NTCP, OP_CLOSE, OP_CONNECT, OP_RECV, OP_SEND, OP_STATE,
-};
-use crate::server::parse_req::Request;
-use crate::server::respond::reply;
-
-pub fn dispatch(sender_pid: u32, req: &Request, body: &[u8], tx: &mut [u8]) {
-    match req.op {
-        OP_CONNECT => connect::handle(sender_pid, req, body, tx),
-        OP_SEND => send::handle(sender_pid, req, body, tx),
-        OP_RECV => recv::handle(sender_pid, req, body, tx),
-        OP_CLOSE => close::handle(sender_pid, req, body, tx),
-        OP_STATE => state::handle(sender_pid, req, body, tx),
-        _ => {
-            let _ = reply(sender_pid, MAGIC_NTCP, req.op, E_BAD_OP, req.request_id, &[], tx);
-        }
-    }
-}
+pub use dispatch::dispatch;
