+++ userland/capsule_net_core/src/server/handlers/tcp/smoltcp_state_to_code.rs
@@ -14,7 +14,20 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-pub mod dispatch;
-pub mod resolve_a;
+use smoltcp::socket::tcp;
 
-pub use dispatch::dispatch;
+pub fn smoltcp_state_to_code(s: tcp::State) -> u8 {
+    match s {
+        tcp::State::Listen => 0,
+        tcp::State::SynSent => 1,
+        tcp::State::SynReceived => 2,
+        tcp::State::Established => 3,
+        tcp::State::CloseWait => 4,
+        tcp::State::FinWait1 => 5,
+        tcp::State::FinWait2 => 6,
+        tcp::State::Closing => 7,
+        tcp::State::TimeWait => 8,
+        tcp::State::LastAck => 9,
+        tcp::State::Closed => 0xFF,
+    }
+}
