+++ userland/capsule_net_core/src/server/handlers/tcp/state.rs
@@ -22,6 +22,8 @@ use crate::server::parse_req::Request;
 use crate::server::respond::reply;
 use crate::state;
 
+use super::smoltcp_state_to_code::smoltcp_state_to_code;
+
 pub fn handle(sender_pid: u32, req: &Request, body: &[u8], tx: &mut [u8]) {
     if body.len() < 4 {
         let _ = reply(sender_pid, MAGIC_NTCP, OP_STATE, E_BAD_LEN, req.request_id, &[], tx);
@@ -42,22 +44,9 @@ pub fn handle(sender_pid: u32, req: &Request, body: &[u8], tx: &mut [u8]) {
         smoltcp_state_to_code(s)
     });
 
-    let code_byte = code.unwrap_or(0xFF);
+    let code_byte = match code {
+        Some(value) => value,
+        None => 0xFF,
+    };
     let _ = reply(sender_pid, MAGIC_NTCP, OP_STATE, E_OK, req.request_id, &[code_byte], tx);
-}
-
-fn smoltcp_state_to_code(s: tcp::State) -> u8 {
-    match s {
-        tcp::State::Listen => 0,
-        tcp::State::SynSent => 1,
-        tcp::State::SynReceived => 2,
-        tcp::State::Established => 3,
-        tcp::State::CloseWait => 4,
-        tcp::State::FinWait1 => 5,
-        tcp::State::FinWait2 => 6,
-        tcp::State::Closing => 7,
-        tcp::State::TimeWait => 8,
-        tcp::State::LastAck => 9,
-        tcp::State::Closed => 0xFF,
-    }
 }
