+++ userland/capsule_net_core/src/server/handlers/udp/mod.rs
@@ -14,23 +14,10 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-mod bind;
-mod recv;
-mod send;
-mod unbind;
+pub mod bind;
+pub mod dispatch;
+pub mod recv;
+pub mod send;
+pub mod unbind;
 
-use crate::protocol::udp::{E_BAD_OP, MAGIC_NUDP, OP_BIND, OP_RECV, OP_SEND, OP_UNBIND};
-use crate::server::parse_req::Request;
-use crate::server::respond::reply;
-
-pub fn dispatch(sender_pid: u32, req: &Request, body: &[u8], tx: &mut [u8]) {
-    match req.op {
-        OP_BIND => bind::handle(sender_pid, req, body, tx),
-        OP_UNBIND => unbind::handle(sender_pid, req, body, tx),
-        OP_SEND => send::handle(sender_pid, req, body, tx),
-        OP_RECV => recv::handle(sender_pid, req, body, tx),
-        _ => {
-            let _ = reply(sender_pid, MAGIC_NUDP, req.op, E_BAD_OP, req.request_id, &[], tx);
-        }
-    }
-}
+pub use dispatch::dispatch;
