--- userland/capsule_net_core/src/state.rs
@@ -1,90 +1,0 @@
-// NONOS Operating System
-// Copyright (C) 2026 NONOS Contributors
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU Affero General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with this program. If not, see <https://www.gnu.org/licenses/>.
-
-use spin::Mutex;
-use smoltcp::iface::{Interface, SocketHandle, SocketSet};
-use crate::device::NicDevice;
-
-pub struct NetState {
-    pub iface: Interface,
-    pub sockets: SocketSet<'static>,
-    pub device: NicDevice,
-    pub dhcp_handle: SocketHandle,
-    pub dns_handle: Option<SocketHandle>,
-}
-
-// SAFETY: The capsule has a single execution context (no threads); spin::Mutex
-// is the exclusive access gate; Interface, SocketSet<'static>, and NicDevice
-// are themselves Send.
-unsafe impl Send for NetState {}
-
-pub static NET: Mutex<Option<NetState>> = Mutex::new(None);
-
-pub struct Lease {
-    pub ip: [u8; 4],
-    pub prefix: u8,
-    pub gw: [u8; 4],
-    pub dns: [u8; 4],
-    pub secs: u32,
-    pub bound: bool,
-}
-
-static LEASE: Mutex<Option<Lease>> = Mutex::new(None);
-
-pub fn lease() -> Option<Lease> {
-    let guard = LEASE.lock();
-    guard.as_ref().map(|l| Lease {
-        ip: l.ip,
-        prefix: l.prefix,
-        gw: l.gw,
-        dns: l.dns,
-        secs: l.secs,
-        bound: l.bound,
-    })
-}
-
-pub fn set_lease(l: Option<Lease>) {
-    *LEASE.lock() = l;
-}
-
-pub fn store(state: NetState) {
-    *NET.lock() = Some(state);
-}
-
-pub fn with_iface<R>(
-    f: impl FnOnce(&mut Interface, &mut SocketSet<'static>, &mut NicDevice) -> R,
-) -> Option<R> {
-    let mut guard = NET.lock();
-    let state = guard.as_mut()?;
-    Some(f(&mut state.iface, &mut state.sockets, &mut state.device))
-}
-
-pub fn with_dns<R>(
-    f: impl FnOnce(&mut Interface, &mut SocketSet<'static>, SocketHandle) -> R,
-) -> Option<R> {
-    let mut guard = NET.lock();
-    let state = guard.as_mut()?;
-    let handle = state.dns_handle?;
-    Some(f(&mut state.iface, &mut state.sockets, handle))
-}
-
-pub fn with_dhcp_and_dns_slot<R>(
-    f: impl FnOnce(&mut Interface, &mut SocketSet<'static>, SocketHandle, &mut Option<SocketHandle>) -> R,
-) -> Option<R> {
-    let mut guard = NET.lock();
-    let state = guard.as_mut()?;
-    Some(f(&mut state.iface, &mut state.sockets, state.dhcp_handle, &mut state.dns_handle))
-}
