+++ userland/capsule_net_core/src/state/mod.rs
@@ -14,7 +14,17 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-pub mod dispatch;
-pub mod resolve_a;
+mod globals;
+mod lease;
+mod store;
+mod types;
+mod with_dhcp_and_dns_slot;
+mod with_dns;
+mod with_iface;
 
-pub use dispatch::dispatch;
+pub use lease::{lease, set_lease};
+pub use store::store;
+pub use types::{Lease, NetState};
+pub use with_dhcp_and_dns_slot::with_dhcp_and_dns_slot;
+pub use with_dns::with_dns;
+pub use with_iface::with_iface;
