+++ userland/capsule_net_core/src/state/store.rs
@@ -14,7 +14,9 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-pub mod dispatch;
-pub mod resolve_a;
+use crate::state::globals::NET;
+use crate::state::types::NetState;
 
-pub use dispatch::dispatch;
+pub fn store(state: NetState) {
+    *NET.lock() = Some(state);
+}
