+++ userland/capsule_net_core/src/state/with_dns.rs
@@ -14,7 +14,15 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-pub mod dispatch;
-pub mod resolve_a;
+use smoltcp::iface::{Interface, SocketHandle, SocketSet};
 
-pub use dispatch::dispatch;
+use crate::state::globals::NET;
+
+pub fn with_dns<R>(
+    f: impl FnOnce(&mut Interface, &mut SocketSet<'static>, SocketHandle) -> R,
+) -> Option<R> {
+    let mut guard = NET.lock();
+    let state = guard.as_mut()?;
+    let handle = state.dns_handle?;
+    Some(f(&mut state.iface, &mut state.sockets, handle))
+}
